+++ NEWS.md
@@ -3,6 +3,7 @@
 # surveydown 0.11.0
 
 - New `yml` argument in `sd_question()`, defaults to `"questions.yml"`. This allows users to utilize yml files for question creation. For details, please refer to the [`questions_yml`](https://surveydown.org/templates/15_questions_yml) template. This enhances the reproducibility of survey design.
+- Update to `sd_create_survey()`: now the The `template` argument is by default `"default"`.
 
 # surveydown 0.10.1